have the bridge subcmd match the rest
